Move internal buffer optimization into PostProcessingPipeline
